Move model related commands to model dir and make model group
ilab now has a model command which has the following subcommands:

ilab model chat
ilab model convert
ilab model serve
ilab model test
ilab model train
ilab model download

ilab model list is new, but simple and I think we should add it in with this redesign to let users get used to the idea of managing their models.

Each child group needs to init their own Lab class due to context issues in click
